+ Add support for (optional) cache dir.
Make oci builder able to use cache dir. If specified, this directory will
contain tar files of already constructed layers, to avoid repeating this process
on subsequent runs.

If cache dir is not specified, the layer tar files will still be constructed in
the (non-mounted) dir, but since they are hard-linked to the build dir, the
overhead is essentially zero.